Transformed APIView to CreateModelMixin and GenericAPIView
