+++ BackEnd/administration/serializers.py
@@ -1,19 +1,61 @@
+from collections import defaultdict
+
+from django.contrib.auth import get_user_model
+from django.core.exceptions import ValidationError
 from rest_framework import serializers
 from authentication.models import CustomUser
 from profiles.models import (
     Profile,
     Region,
 )
+from utils.administration.create_password import generate_password
+from utils.administration.send_email import send_email_about_admin_registration
+from validation.validate_password import validate_password_long
 from .models import AutoModeration, ModerationEmail
 
+User = get_user_model()
 
-class AdminRegionSerialaizer(serializers.ModelSerializer):
+
+class AdminRegionSerializer(serializers.ModelSerializer):
     class Meta:
         model = Region
         fields = (
             "id",
             "name_ukr",
         )
+
+
+class AdminRegistrationSerializer(serializers.Serializer):
+    email = serializers.EmailField(
+        write_only=True,
+    )
+
+    def validate(self, value):
+        custom_errors = defaultdict(list)
+        email = value.get("email").lower()
+
+        if User.objects.filter(email=email).exists():
+            custom_errors["email"].append("Email is already registered")
+
+        if custom_errors:
+            raise serializers.ValidationError(custom_errors)
+        return value
+
+    def create(self, validated_data):
+        email = validated_data.get("email")
+        password = generate_password()
+        name = "admin"
+        surname = "admin"
+        admin = User.objects.create(
+            email=email,
+            name=name,
+            surname=surname,
+            password=password,
+
+        )
+        admin.save()
+        send_email_about_admin_registration(email, password)
+        return admin
 
 
 class AdminUserListSerializer(serializers.ModelSerializer):
@@ -48,7 +90,7 @@ class AdminUserDetailSerializer(serializers.ModelSerializer):
 
 class AdminCompanyListSerializer(serializers.ModelSerializer):
     person = AdminUserDetailSerializer(read_only=True)
-    regions = AdminRegionSerialaizer(many=True, read_only=True)
+    regions = AdminRegionSerializer(many=True, read_only=True)
 
     class Meta:
         model = Profile
@@ -76,7 +118,7 @@ class AdminCompanyDetailSerializer(serializers.ModelSerializer):
     activities = serializers.SlugRelatedField(
         many=True, slug_field="name", read_only=True
     )
-    regions = AdminRegionSerialaizer(many=True, read_only=True)
+    regions = AdminRegionSerializer(many=True, read_only=True)
     banner_image = serializers.ImageField(
         source="banner.image_path", required=False
     )
